third_party/blink/web_tests/platform/linux/external/wpt/speculation-rules/prerender/media-autoplay-expected.txt

This is a testharness.js-based test.
[FAIL] media autoplay should defer playaback
  assert_equals: expected false but got true
Harness: the test ran to completion.
